add authorization to delete and tests

+++ apps/web/src/server/api/routers/model.ts
@@ -3,20 +3,9 @@ import { z } from "zod";
 
 import { createTRPCRouter, publicProcedure, protectedProcedure } from "../trpc";
 import { fromUint8Array } from "js-base64";
+import { TRPCError } from "@trpc/server";
 
-// want to allow a user to create a document without signing in
-// this document shouldn't be saved
-
-// const toBase64 = (u8: Uint8Array) => Buffer.from(u8).toString("base64");
-// const fromBase64 = (b64: string) => new Uint8Array(Buffer.from(b64, "base64"));
-
-/**
- * Since there's currently no server,
- * create a ydoc base state so all updates are applied on top of this document and can be synced
- * this must be consistent across clients (with a consistent clientId) - generated once
- * any updates to the intial shared state should extend this template
- * https://discuss.yjs.dev/t/merging-changes-from-one-document-into-another/499/4
- */
+// create base document
 const ydoc = new Y.Doc();
 const ymap = ydoc.getMap("cubes");
 const layersMap = new Y.Map();
@@ -29,9 +18,6 @@ ymap.set("layers", layersMap);
 const layerOrder = new Y.Array();
 layerOrder.insert(0, ["layer_1"]);
 ymap.set("layerOrder", layerOrder);
-// const template = fromUint8Array(Y.encodeStateAsUpdate(ydoc));
-// const template =
-//   "AQfq7dfVBQAnAQVjdWJlcwVjdWJlcwEnAQVjdWJlcwZsYXllcnMBJwDq7dfVBQEHbGF5ZXJfMQEoAOrt19UFAgRuYW1lAXcHTGF5ZXIgMSgA6u3X1QUCB3Zpc2libGUBeCcBBWN1YmVzCmxheWVyT3JkZXIACADq7dfVBQUBdwdsYXllcl8xAA==";
 
 export const modelRouter = createTRPCRouter({
   create: protectedProcedure.mutation(({ ctx }) => {
@@ -71,7 +57,22 @@ export const modelRouter = createTRPCRouter({
 
   delete: protectedProcedure
     .input(z.object({ id: z.string() }))
-    .mutation(({ input, ctx }) => {
+    .mutation(async ({ input, ctx }) => {
+      const model = await ctx.prisma.model.findUnique({
+        where: { id: input.id },
+      });
+
+      if (!model) {
+        throw new TRPCError({ code: "NOT_FOUND" });
+      }
+
+      if (model.authorId !== ctx.session.user.id) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "Only owner of model may delete.",
+        });
+      }
+
       return ctx.prisma.model.delete({ where: { id: input.id } });
     }),
 });
